pop out side search bar as component
